perbaiki fitur

+++ database/migrations/2022_04_02_082807_create_pengeluarans_table.php
@@ -16,7 +16,7 @@ class CreatePengeluaransTable extends Migration
     {
         Schema::create('pengeluarans', function (Blueprint $table) {
             $table->id();
-            $table->float('nominal');
+            $table->double('nominal');
             $table->text('keterangan');
             $table->timestamps();
         });
